+++ CHANGES.rst
@@ -6,7 +6,9 @@ Changelog
 
 New:
 
-- *add item here*
+- Support chameleon attributes tal:switch and tal:case.
+  Fixes issue #24.
+  [ale-rt]
 
 Fixes: